server_fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
